feat: Get group todos

+++ controllers/groupController.js
@@ -1,5 +1,27 @@
+const mongoose = require("../loader/mongoose");
 const Group = require("../models/Group");
 const catchAsync = require("../utils/catchAsync");
+
+exports.getGroupTodos = catchAsync(async (req, res, next) => {
+  const { groupId } = req.params;
+
+  if (!mongoose.isValidObjectId(groupId)) {
+    return res.json({
+      result: "error",
+      error: {
+        message: "유효하지 않은 그룹입니다.",
+        status: 400,
+      },
+    });
+  }
+
+  const group = await Group.findById(groupId);
+
+  return res.json({
+    result: "success",
+    data: group.todos,
+  });
+});
 
 exports.createGroup = catchAsync(async (req, res, next) => {
   const { title } = req.body;
@@ -23,8 +45,18 @@ exports.createGroup = catchAsync(async (req, res, next) => {
 });
 
 exports.updateGroup = catchAsync(async (req, res, next) => {
-  const groupId = req.params.id;
+  const { groupId } = req.params;
   const { title } = req.body;
+
+  if (!mongoose.isValidObjectId(groupId)) {
+    return res.json({
+      result: "error",
+      error: {
+        message: "유효하지 않은 그룹입니다.",
+        status: 400,
+      },
+    });
+  }
 
   if (!title) {
     return res.json({
@@ -42,7 +74,19 @@ exports.updateGroup = catchAsync(async (req, res, next) => {
 });
 
 exports.deleteGroup = catchAsync(async (req, res, next) => {
-  const groupId = req.params.id;
+  const { groupId } = req.params;
+
+  if (!mongoose.isValidObjectId(groupId)) {
+    return res.json({
+      result: "error",
+      error: {
+        message: "유효하지 않은 그룹입니다.",
+        status: 400,
+      },
+    });
+  }
 
   await Group.findByIdAndDelete(groupId);
+
+  return res.json({ result: "success" });
 });
